```
name         old time/op  new time/op  delta
AddSafe-8    96.6µs ± 1%  83.1µs ± 1%  -14.03%  (p=0.000 n=10+10)
AddUnsafe-8  78.4µs ± 1%  67.7µs ± 2%  -13.57%  (p=0.000 n=9+10)
```