@deckarep 2022-03-29
@orisano - this is a slick improvement and clever. I wouldn't have expected to realize such gains.

I suppose the overhead is in getting rid of the need to do a comparison/branch.